organize course table card related files to folder
